Rename Mongo adapters / tests to MongoDB
