+++ src/Database/MongoDB.php
@@ -15,10 +15,12 @@ use Imbo\Model\Image,
     Imbo\Resource\Images\Query,
     Imbo\Exception\DatabaseException,
     Imbo\Exception\DuplicateImageIdentifierException,
-    MongoClient,
-    MongoCollection,
-    MongoException,
-    MongoDuplicateKeyException,
+    Imbo\Helpers\BSONToArray,
+    MongoDB\Client as MongoClient,
+    MongoDB\Driver\Manager,
+    MongoDB\Driver\Command,
+    MongoDB\Collection as MongoCollection,
+    MongoDB\Driver\Exception\Exception as MongoException,
     DateTime,
     DateTimeZone;
 
@@ -32,7 +34,7 @@ use Imbo\Model\Image,
  * - (string) databaseName Name of the database. Defaults to 'imbo'
  * - (string) server The server string to use when connecting to MongoDB. Defaults to
  *                   'mongodb://localhost:27017'
- * - (array) options Options to use when creating the MongoClient instance. Defaults to
+ * - (array) options Options to use when creating the MongoDB\Client instance. Defaults to
  *                   ['connect' => true, 'connectTimeoutMS' => 1000].
  *
  * @author Christer Edvartsen <cogo@starzinger.net>
@@ -40,7 +42,7 @@ use Imbo\Model\Image,
  */
 class MongoDB implements DatabaseInterface {
     /**
-     * Mongo client instance
+     * MongoDB client instance
      *
      * @var MongoClient
      */
@@ -71,14 +73,22 @@ class MongoDB implements DatabaseInterface {
     ];
 
     /**
+     * BSONToArray helper
+     *
+     * @var BSONToArray
+     */
+    private $bsonToArray;
+
+    /**
      * Class constructor
      *
      * @param array $params Parameters for the driver
      * @param MongoClient $client MongoClient instance
      * @param MongoCollection $imageCollection MongoCollection instance for the images
      * @param MongoCollection $shortUrlCollection MongoCollection instance for the short URLs
+     * @param BSONToArray $bsonToArray Helper to recursively convert documents to arrays
      */
-    public function __construct(array $params = null, MongoClient $client = null, MongoCollection $imageCollection = null, MongoCollection $shortUrlCollection = null) {
+    public function __construct(array $params = null, MongoClient $client = null, MongoCollection $imageCollection = null, MongoCollection $shortUrlCollection = null, BSONToArray $bsonToArray = null) {
         if ($params !== null) {
             $this->params = array_replace_recursive($this->params, $params);
         }
@@ -94,6 +104,12 @@ class MongoDB implements DatabaseInterface {
         if ($shortUrlCollection !== null) {
             $this->collections['shortUrl'] = $shortUrlCollection;
         }
+
+        if ($bsonToArray === null) {
+            $bsonToArray = new BSONToArray();
+        }
+
+        $this->bsonToArray = $bsonToArray;
     }
 
     /**
@@ -112,10 +128,9 @@ class MongoDB implements DatabaseInterface {
 
         if ($updateIfDuplicate && $this->imageExists($user, $imageIdentifier)) {
             try {
-                $this->getImageCollection()->update(
+                $this->getImageCollection()->updateOne(
                     ['user' => $user, 'imageIdentifier' => $imageIdentifier],
-                    ['$set' => ['updated' => $now]],
-                    ['multiple' => false]
+                    ['$set' => ['updated' => $now]]
                 );
 
                 return true;
@@ -125,25 +140,32 @@ class MongoDB implements DatabaseInterface {
         }
 
         $data = [
-            'size' => $image->getFilesize(),
-            'user' => $user,
-            'imageIdentifier' => $imageIdentifier,
-            'extension' => $image->getExtension(),
-            'mime' => $image->getMimeType(),
-            'metadata' => [],
-            'added' => $added ?: $now,
-            'updated' => $updated ?: $now,
-            'width' => $image->getWidth(),
-            'height' => $image->getHeight(),
-            'checksum' => $image->getChecksum(),
+            'size'             => $image->getFilesize(),
+            'user'             => $user,
+            'imageIdentifier'  => $imageIdentifier,
+            'extension'        => $image->getExtension(),
+            'mime'             => $image->getMimeType(),
+            'metadata'         => [],
+            'added'            => $added ?: $now,
+            'updated'          => $updated ?: $now,
+            'width'            => $image->getWidth(),
+            'height'           => $image->getHeight(),
+            'checksum'         => $image->getChecksum(),
             'originalChecksum' => $image->getOriginalChecksum(),
         ];
 
         try {
-            $this->getImageCollection()->insert($data);
-        } catch (MongoDuplicateKeyException $e) {
-            throw new DuplicateImageIdentifierException('Duplicate image identifier when attempting to insert image into DB.', 503);
+            $this->getImageCollection()->insertOne($data);
         } catch (MongoException $e) {
+            foreach ($e->getWriteResult()->getWriteErrors() as $error) {
+                if ($error->getCode() === 11000) {
+                    throw new DuplicateImageIdentifierException(
+                        'Duplicate image identifier when attempting to insert image into DB.',
+                        503
+                    );
+                }
+            }
+
             throw new DatabaseException('Unable to save image data', 500, $e);
         }
 
@@ -164,10 +186,10 @@ class MongoDB implements DatabaseInterface {
                 throw new DatabaseException('Image not found', 404);
             }
 
-            $this->getImageCollection()->remove(
-                ['user' => $user, 'imageIdentifier' => $imageIdentifier],
-                ['justOne' => true]
-            );
+            $this->getImageCollection()->deleteOne([
+                'user' => $user,
+                'imageIdentifier' => $imageIdentifier,
+            ]);
         } catch (MongoException $e) {
             throw new DatabaseException('Unable to delete image data', 500, $e);
         }
@@ -184,10 +206,9 @@ class MongoDB implements DatabaseInterface {
             $existing = $this->getMetadata($user, $imageIdentifier);
             $updatedMetadata = array_merge($existing, $metadata);
 
-            $this->getImageCollection()->update(
+            $this->getImageCollection()->updateOne(
                 ['user' => $user, 'imageIdentifier' => $imageIdentifier],
-                ['$set' => ['updated' => time(), 'metadata' => $updatedMetadata]],
-                ['multiple' => false]
+                ['$set' => ['updated' => time(), 'metadata' => $updatedMetadata]]
             );
         } catch (MongoException $e) {
             throw new DatabaseException('Unable to update meta data', 500, $e);
@@ -213,7 +234,7 @@ class MongoDB implements DatabaseInterface {
             throw new DatabaseException('Image not found', 404);
         }
 
-        return isset($data['metadata']) ? $data['metadata'] : [];
+        return $this->bsonToArray->toArray($data['metadata']->getArrayCopy());
     }
 
     /**
@@ -230,10 +251,9 @@ class MongoDB implements DatabaseInterface {
                 throw new DatabaseException('Image not found', 404);
             }
 
-            $this->getImageCollection()->update(
+            $this->getImageCollection()->updateOne(
                 ['user' => $user, 'imageIdentifier' => $imageIdentifier],
-                ['$set' => ['metadata' => []]],
-                ['multiple' => false]
+                ['$set' => ['metadata' => []]]
             );
         } catch (MongoException $e) {
             throw new DatabaseException('Unable to delete meta data', 500, $e);
@@ -309,25 +329,29 @@ class MongoDB implements DatabaseInterface {
         }
 
         try {
-            $cursor = $this->getImageCollection()->find($queryData, $fields)
-                                                 ->limit($query->limit())
-                                                 ->sort($sort);
+            $options = [
+                'projection' => $fields,
+                'limit' => $query->limit(),
+                'sort' => $sort,
+            ];
 
             // Skip some images if a page has been set
             if (($page = $query->page()) > 1) {
                 $skip = $query->limit() * ($page - 1);
-                $cursor->skip($skip);
+                $options['skip'] = $skip;
             }
+
+            $cursor = $this->getImageCollection()->find($queryData, $options);
 
             foreach ($cursor as $image) {
                 unset($image['_id']);
                 $image['added'] = new DateTime('@' . $image['added'], new DateTimeZone('UTC'));
                 $image['updated'] = new DateTime('@' . $image['updated'], new DateTimeZone('UTC'));
-                $images[] = $image;
+                $images[] = $this->bsonToArray->toArray($image->getArrayCopy());
             }
 
             // Update model
-            $model->setHits($cursor->count());
+            $model->setHits($this->getImageCollection()->count($queryData));
         } catch (MongoException $e) {
             throw new DatabaseException('Unable to search for images', 500, $e);
         }
@@ -383,15 +407,15 @@ class MongoDB implements DatabaseInterface {
                 $query['imageIdentifier'] = $imageIdentifier;
             }
 
-            // Create the cursor
-            $cursor = $this->getImageCollection()->find($query, ['updated' => true])
-                                                 ->limit(1)
-                                                 ->sort([
-                                                     'updated' => -1,
-                                                 ]);
-
-            // Fetch the next row
-            $data = $cursor->getNext();
+            // Find a document
+            $data = $this->getImageCollection()->findOne($query, [
+                'sort' => [
+                    'updated' => -1,
+                ],
+                'projection' => [
+                    'updated' => true
+                ],
+            ]);
         } catch (MongoException $e) {
             throw new DatabaseException('Unable to fetch image data', 500, $e);
         }
@@ -429,22 +453,32 @@ class MongoDB implements DatabaseInterface {
      */
     public function getNumBytes($user = null) {
         try {
-            $collection = $this->getImageCollection();
-            $group = ['$group' => ['_id' => null, 'numBytes' => ['$sum' => '$size']]];
+            $pipeline = [
+                [
+                    '$group' => [
+                        '_id' => null,
+                        'numBytes' => [
+                            '$sum' => '$size',
+                        ],
+                    ],
+                ],
+            ];
 
             if ($user) {
-                $results = $collection->aggregate(['$match' => ['user' => $user]], $group);
-            } else {
-                $results = $collection->aggregate($group);
+                array_unshift($pipeline, [
+                    '$match' => [
+                        'user' => $user,
+                    ],
+                ]);
             }
 
-            $result = $results['result'];
+            $result = $this->getImageCollection()->aggregate($pipeline, ['useCursor' => false]);
 
-            if (empty($result)) {
+            if (!count($result)) {
                 return 0;
             }
 
-            return (int) $result[0]['numBytes'];
+            return (int) $result[0]->numBytes;
         } catch (MongoException $e) {
             throw new DatabaseException('Unable to fetch information from the database', 500, $e);
         }
@@ -468,8 +502,12 @@ class MongoDB implements DatabaseInterface {
      */
     public function getStatus() {
         try {
-            return $this->getMongoClient()->connect();
-        } catch (DatabaseException $e) {
+            // Create a manager and try to get servers
+            $manager = new Manager($this->params['server']);
+            $manager->executeCommand($this->params['databaseName'], new Command(['ping' => 1]));
+
+            return true;
+        } catch (MongoException $e) {
             return false;
         }
     }
@@ -519,7 +557,7 @@ class MongoDB implements DatabaseInterface {
                 'query' => serialize($query),
             ];
 
-            $this->getShortUrlCollection()->insert($data);
+            $this->getShortUrlCollection()->insertOne($data);
         } catch (MongoException $e) {
             throw new DatabaseException('Unable to create short URL', 500, $e);
         }
@@ -588,7 +626,7 @@ class MongoDB implements DatabaseInterface {
         }
 
         try {
-            $this->getShortUrlCollection()->remove($query);
+            $this->getShortUrlCollection()->deleteMany($query);
         } catch (MongoException $e) {
             throw new DatabaseException('Unable to delete short URLs', 500, $e);
         }
